pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mlip-phonons"
version = "0.1.0"
description = "Phonon workflow with MLIP calculators"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "ase",
  "numpy",
  "pyyaml",
  "phonopy",
  "matplotlib",
  "torch",
]

[project.optional-dependencies]
mace = ["mace-torch"]         
mattersim = ["mattersim"]
vdw = ["dftd3"]
orb = ["orb-models", "pynanoflann@git+https://github.com/dwastberg/pynanoflann#egg=af434039ae14bedcbb838a7808924d6689274168"]
pet = ["upet"]
matgl = ["matgl", "dgl"]

[project.scripts]
mlip-phonons = "mlip_phonons.main:main"
mlip-neb = "NEB.run_neb_raw_v2:main"
mlip-coup = "coupling_modes.phonon_coupling:main"

[tool.setuptools]
package-dir = {"" = "src"}
packages = ["mlip_phonons", "mlip_phonons.tools", "NEB", "NEB.neb_tools", "coupling_modes.coup_tools"]
include-package-data = true
